moved explored functionality into gameworld
