+++ jam.TODO
@@ -5,7 +5,6 @@ Blocked:
     ✔ Blob Keep https://itch.io/jam/dcjam2021/rate/978972 @done(21-04-08 04:10)
     needs a greater time investment
     ✔ The Ghostly LAB-yrinth https://itch.io/jam/dcjam2021/rate/986339 @done(21-04-08 04:36)
-    ✘ Mystidinia https://itch.io/jam/dcjam2021/rate/986114 @cancelled(21-04-08 04:24)
     ✔ Hand 2 Hand https://itch.io/jam/dcjam2021/rate/986391 @done(21-04-08 04:36)
     ✔ EFCC: Escape From Cosmic Cube https://itch.io/jam/dcjam2021/rate/986296 @done(21-04-08 05:03)
     ✔ From A Lost Tunnel https://itch.io/jam/dcjam2021/rate/986546 @done(21-04-08 04:42)
@@ -21,9 +20,10 @@ Blocked:
     ✔ Space Crawlers #dcjam2021 https://itch.io/jam/dcjam2021/rate/986649 @done(21-04-08 05:51)
     ✔ The Infinite Grimoire https://itch.io/jam/dcjam2021/rate/986377 @done(21-04-08 05:58)
         need winrar
-Jam Games To Review:
+20210409(planned... we'll see how it actually goes):
     ☐ The Reign of the Lizard Wizard https://itch.io/jam/dcjam2021/rate/986676
         need winrar
+    ☐ Mystidinia https://itch.io/jam/dcjam2021/rate/986114
     ☐ Calderan Chronicles - Prologue https://itch.io/jam/dcjam2021/rate/987105
     ☐ The Dark Forest in Heaven https://itch.io/jam/dcjam2021/rate/987111
     ☐ Synthyfriends Dungeon https://itch.io/jam/dcjam2021/rate/987118
@@ -34,6 +34,7 @@ Jam Games To Review:
     ☐ Goblin Slaying Dungeon Crawler https://itch.io/jam/dcjam2021/rate/987400
     ☐ Two Problems: A Regular Space Adventure https://itch.io/jam/dcjam2021/rate/987164
     ☐ This Mall is a Maze and I Gotta Pee https://itch.io/jam/dcjam2021/rate/986502
+Jam Games To Review:
     ☐ Hotel Inverse https://itch.io/jam/dcjam2021/rate/986943
     ☐ Prototype dungeon crawler https://itch.io/jam/dcjam2021/rate/987495
     ☐ Ankehotep's Prison https://itch.io/jam/dcjam2021/rate/987063
